separated top Artists and working fine
